# packages/core/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ScoutGlobe API",
    "description": "Futbolcu verisi, transfer onerisi ve future-star kesfi icin API.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/leagues": {
      "get": {
        "tags": [
          "leagues"
        ],
        "summary": "Ligleri listele",
        "operationId": "list_leagues_leagues_get",
        "parameters": [
          {
            "name": "country",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "minLength": 2,
                  "maxLength": 2
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO ulke kodu",
              "title": "Country"
            },
            "description": "ISO ulke kodu"
          },
          {
            "name": "tier",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 10,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LeagueOut"
                  },
                  "title": "Response List Leagues Leagues Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/leagues/{league_id}": {
      "get": {
        "tags": [
          "leagues"
        ],
        "summary": "Lig detayi ve kulupleri",
        "operationId": "get_league_leagues__league_id__get",
        "parameters": [
          {
            "name": "league_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "League Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeagueDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClubSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "leagueId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Leagueid"
          },
          "logoUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logourl"
          },
          "squadSize": {
            "type": "integer",
            "title": "Squadsize",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "name"
        ],
        "title": "ClubSummary"
      },
      "CountryOut": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "nameTr": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nametr"
          },
          "lat": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lat"
          },
          "lng": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lng"
          }
        },
        "type": "object",
        "required": [
          "code",
          "name"
        ],
        "title": "CountryOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LeagueDetail": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "countryCode": {
            "type": "string",
            "title": "Countrycode"
          },
          "logoUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logourl"
          },
          "tier": {
            "type": "integer",
            "title": "Tier"
          },
          "strengthCoef": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strengthcoef"
          },
          "season": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Season"
          },
          "clubCount": {
            "type": "integer",
            "title": "Clubcount",
            "default": 0
          },
          "playerCount": {
            "type": "integer",
            "title": "Playercount",
            "default": 0
          },
          "country": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CountryOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "squadSeason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Squadseason"
          },
          "clubs": {
            "items": {
              "$ref": "#/components/schemas/ClubSummary"
            },
            "type": "array",
            "title": "Clubs",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "countryCode",
          "tier"
        ],
        "title": "LeagueDetail"
      },
      "LeagueOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "countryCode": {
            "type": "string",
            "title": "Countrycode"
          },
          "logoUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logourl"
          },
          "tier": {
            "type": "integer",
            "title": "Tier"
          },
          "strengthCoef": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strengthcoef"
          },
          "season": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Season"
          },
          "clubCount": {
            "type": "integer",
            "title": "Clubcount",
            "default": 0
          },
          "playerCount": {
            "type": "integer",
            "title": "Playercount",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "countryCode",
          "tier"
        ],
        "title": "LeagueOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}